app/requirements.txt
requests
flask
keras
matplotlib
numpy
pandas
image
h5py
grpcio-tools
mysql-connector